styles
зщзділила на декілька функцій те що написав Олег, та додала стилі
