+++ src/js/library-card/date-optimizer.js
@@ -1,0 +1,7 @@
+export default function dateOptimizer(releaseDate, firstDate) {
+    if (releaseDate) {
+      return releaseDate.slice(0, 4);
+    }
+  
+    return firstDate.slice(0, 4);
+  }
